1.18.2 backport
All of the new features have been backported to 1.18.2. The reason for this is that 1.18.2 has functioning worldgen, while 1.19 does not. This means that in this version, you should be able to play the entire mod from the start (earth) to the finish (glacio) in survival.

+++ src/main/java/com/github/alexnijjar/beyond_earth/entities/vehicles/VehicleEntity.java
@@ -2,10 +2,18 @@ package com.github.alexnijjar.beyond_earth.entities.vehicles;
 
 import javax.annotation.Nullable;
 
+import com.github.alexnijjar.beyond_earth.BeyondEarth;
+import com.github.alexnijjar.beyond_earth.gui.VehicleScreenHandlerFactory;
+import com.github.alexnijjar.beyond_earth.items.vehicles.VehicleItem;
+import com.github.alexnijjar.beyond_earth.registry.ModFluids;
+import com.github.alexnijjar.beyond_earth.util.CustomInventory;
+import com.github.alexnijjar.beyond_earth.util.FluidUtils;
 import com.github.alexnijjar.beyond_earth.util.ModUtils;
 
-import net.minecraft.block.Blocks;
-import net.minecraft.block.PowderSnowBlock;
+import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
+import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
+import net.fabricmc.fabric.api.transfer.v1.storage.base.SingleVariantStorage;
+import net.fabricmc.fabric.api.transfer.v1.transaction.Transaction;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ItemEntity;
@@ -17,91 +25,162 @@ import net.minecraft.entity.data.TrackedDataHandlerRegistry;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtElement;
 import net.minecraft.network.Packet;
 import net.minecraft.network.packet.s2c.play.EntitySpawnS2CPacket;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.registry.Registry;
+import net.minecraft.world.GameRules;
 import net.minecraft.world.World;
 import net.minecraft.world.explosion.Explosion;
 
-public class VehicleEntity extends Entity {
+public abstract class VehicleEntity extends Entity {
 
-    protected int clientInterpolationSteps;
     protected double clientX;
     protected double clientY;
     protected double clientZ;
     public double clientYaw;
     public double clientPitch;
+    private int field_7708;
+
     protected double clientXVelocity;
     protected double clientYVelocity;
     protected double clientZVelocity;
 
+    public float previousYaw;
+
+    public final SingleVariantStorage<FluidVariant> inputTank = FluidUtils.createTank(this.getTankSize());
+    private final CustomInventory inventory = new CustomInventory(this.getInventorySize());
+
     protected static final TrackedData<Float> SPEED = DataTracker.registerData(VehicleEntity.class, TrackedDataHandlerRegistry.FLOAT);
+    public static final TrackedData<Integer> FLUID_AMOUNT = DataTracker.registerData(VehicleEntity.class, TrackedDataHandlerRegistry.INTEGER);
+    public static final TrackedData<String> FLUID_VARIANT = DataTracker.registerData(VehicleEntity.class, TrackedDataHandlerRegistry.STRING);
 
     public VehicleEntity(EntityType<?> type, World world) {
         super(type, world);
-        this.syncClient();
     }
 
     @Override
     protected void initDataTracker() {
         this.dataTracker.startTracking(SPEED, 0.0f);
+        this.dataTracker.startTracking(FLUID_AMOUNT, 0);
+        this.dataTracker.startTracking(FLUID_VARIANT, "beyond_earth:fuel");
     }
 
     @Override
     protected void readCustomDataFromNbt(NbtCompound nbt) {
+        this.inventory.readNbtList(nbt.getList("Inventory", NbtElement.COMPOUND_TYPE));
+        inputTank.variant = FluidVariant.fromNbt(nbt.getCompound("inputFluid"));
+        inputTank.amount = nbt.getLong("inputAmount");
     }
 
     @Override
     protected void writeCustomDataToNbt(NbtCompound nbt) {
+        nbt.put("Inventory", this.inventory.toNbtList());
+        nbt.put("inputFluid", inputTank.variant.toNbt());
+        nbt.putLong("inputAmount", inputTank.amount);
     }
 
     @Override
-    public Packet<?> createSpawnPacket() {
-        return new EntitySpawnS2CPacket(this);
+    public void updateTrackedPositionAndAngles(double x, double y, double z, float yaw, float pitch, int interpolationSteps, boolean interpolate) {
+        this.clientX = x;
+        this.clientY = y;
+        this.clientZ = z;
+        this.clientYaw = yaw;
+        this.clientPitch = pitch;
+        this.field_7708 = 10;
+        this.setVelocity(this.clientXVelocity, this.clientYVelocity, this.clientZVelocity);
+    }
+
+    private void updatePositionAndRotation() {
+        if (this.isLogicalSideForUpdatingMovement()) {
+            this.field_7708 = 0;
+            this.updateTrackedPosition(this.getX(), this.getY(), this.getZ());
+        }
+        if (this.field_7708 <= 0) {
+            return;
+        }
+        double d = this.getX() + (this.clientX - this.getX()) / (double) this.field_7708;
+        double e = this.getY() + (this.clientY - this.getY()) / (double) this.field_7708;
+        double f = this.getZ() + (this.clientZ - this.getZ()) / (double) this.field_7708;
+        double g = MathHelper.wrapDegrees(this.clientYaw - (double) this.getYaw());
+        this.setYaw(this.getYaw() + (float) g / (float) this.field_7708);
+        this.setPitch(this.getPitch() + (float) (this.clientPitch - (double) this.getPitch()) / (float) this.field_7708);
+        --this.field_7708;
+        this.setPosition(d, e, f);
+        this.setRotation(this.getYaw(), this.getPitch());
     }
 
     @Override
-    public boolean collides() {
-        return true;
-    }
-
-    @Override
-    public boolean isCollidable() {
-        return true;
+    public void setVelocityClient(double x, double y, double z) {
+        this.clientXVelocity = x;
+        this.clientYVelocity = y;
+        this.clientZVelocity = z;
+        this.setVelocity(this.clientXVelocity, this.clientYVelocity, this.clientZVelocity);
     }
 
     @Override
     public void tick() {
-        super.tick();
+        this.previousYaw = this.getYaw();
 
-        this.syncClient();
-        this.checkBlockCollision();
+        super.tick();
+        this.updatePositionAndRotation();
+        this.doMovement();
+        this.slowDown();
         this.doGravity();
+        this.move(MovementType.SELF, this.getVelocity());
+        this.checkBlockCollision();
+
+        this.tryInsertingIntoTank();
+        if (!this.world.isClient) {
+            this.dataTracker.set(FLUID_AMOUNT, (int) this.inputTank.amount);
+            this.dataTracker.set(FLUID_VARIANT, Registry.FLUID.getId(this.inputTank.variant.getFluid()).toString());
+        }
     }
 
-    public void syncClient() {
-        if (this.world.isClient) {
-            if (this.clientInterpolationSteps > 0) {
-                double d = this.getX() + (this.clientX - this.getX()) / (double) this.clientInterpolationSteps;
-                double e = this.getY() + (this.clientY - this.getY()) / (double) this.clientInterpolationSteps;
-                double f = this.getZ() + (this.clientZ - this.getZ()) / (double) this.clientInterpolationSteps;
-                double g = MathHelper.wrapDegrees(this.clientYaw - (double) this.getYaw());
-                this.setYaw(this.getYaw() + (float) g / (float) this.clientInterpolationSteps);
-                this.setPitch(this.getPitch() + (float) (this.clientPitch - (double) this.getPitch()) / (float) this.clientInterpolationSteps);
-                --this.clientInterpolationSteps;
-                this.setPosition(d, e, f);
-                this.setRotation(this.getYaw(), this.getPitch());
-            } else {
-                this.refreshPosition();
-                this.setRotation(this.getYaw(), this.getPitch());
+    // Sets the velocity based on the current speed and the current direction.
+    public void doMovement() {
+        this.setPitch(0);
+        Vec3d movement = this.getRotationVector(this.getPitch(), this.getYaw());
+
+        // Save the current y velocity so we can use it later.
+        double yVelocity = this.getVelocity().getY();
+
+        this.setVelocity(this.getVelocity().add(movement.getX(), 0.0, movement.getZ()).multiply(this.getSpeed()));
+
+        // Prevent the vehicle from going too fast.
+        double maxVelocity = 0.75;
+        if (this.getVelocity().lengthSquared() > maxVelocity * maxVelocity) {
+            this.setVelocity(this.getVelocity().normalize().multiply(maxVelocity));
+        }
+
+        // Set the y velocity back to the original value, as it was modified by the movement.
+        this.setVelocity(new Vec3d(this.getVelocity().getX(), yVelocity, this.getVelocity().getZ()));
+    }
+
+    // Slow down the vehicle until a full stop is reached.
+    public void slowDown() {
+        this.setSpeed(this.getSpeed() / 1.1f);
+        if (this.getSpeed() < 0.001 && this.getSpeed() > -0.001) {
+            this.setSpeed(0.0f);
+        }
+        this.setSpeed(MathHelper.clamp(this.getSpeed(), BeyondEarth.CONFIG.vehicles.minSpeed, BeyondEarth.CONFIG.vehicles.maxSpeed));
+    }
+
+    // Apply gravity to the vehicle.
+    public void doGravity() {
+        if (!this.hasNoGravity()) {
+            this.setVelocity(this.getVelocity().add(0, -0.03, 0));
+            if (this.getVelocity().getY() < BeyondEarth.CONFIG.vehicles.gravity) {
+                this.setVelocity(new Vec3d(this.getVelocity().getX(), BeyondEarth.CONFIG.vehicles.gravity, this.getVelocity().getZ()));
             }
-            return;
         }
     }
 
@@ -114,74 +193,14 @@ public class VehicleEntity extends Entity {
     }
 
     @Override
-    public void updateTrackedPositionAndAngles(double x, double y, double z, float yaw, float pitch, int interpolationSteps, boolean interpolate) {
-        this.clientX = x;
-        this.clientY = y;
-        this.clientZ = z;
-        this.clientYaw = yaw;
-        this.clientPitch = pitch;
-        this.clientInterpolationSteps = interpolationSteps + 2;
-        this.setVelocity(this.clientXVelocity, this.clientYVelocity, this.clientZVelocity);
-    }
-
-    @Override
-    public void setVelocityClient(double x, double y, double z) {
-        this.clientXVelocity = x;
-        this.clientYVelocity = y;
-        this.clientZVelocity = z;
-        this.setVelocity(this.clientXVelocity, this.clientYVelocity, this.clientZVelocity);
-    }
-
-    public void setAnglesClient(double pitch, double yaw) {
-        this.clientYaw = yaw;
-        this.clientPitch = pitch;
-    }
-
-    public void doGravity() {
-        if (!this.hasNoGravity()) {
-            double modifier = 1.0;
-            if (this.submergedInWater) {
-                modifier *= 0.01;
-            }
-
-            float gravity = ModUtils.getPlanetGravity(this.world);
-            if (this instanceof LanderEntity) {
-                modifier *= MathHelper.clamp(gravity * 2, 0.2f, 1.0f);
-            } else {
-                modifier *= gravity;
-            }
-
-            this.setVelocity(this.getVelocity().add(0.0, -0.04 * modifier, 0.0));
-        }
-
-        // Slow down the vehicle.
-        
-        this.setSpeed(this.getSpeed() / 1.1f);
-        if (this.getSpeed() < 0.001 && this.getSpeed() > -0.001) {
-            this.setSpeed(0.0f);
-        }
-        this.setSpeed(MathHelper.clamp(this.getSpeed(), -0.1f, 0.2f));
-
-        
-        // Move the vehicle.
-        BlockPos velocityAffectingPos = this.getVelocityAffectingPos();
-        float slipperiness = this.world.getBlockState(velocityAffectingPos).getBlock().getSlipperiness();
-        Vec3d movement = this.applyMovementInput(new Vec3d(this.getPitch(), 0, 1), slipperiness);
-        double speedModifier = this.onGround ? slipperiness * 0.80f : 0.80f;
-        this.setVelocity(movement.getX() * speedModifier, movement.getY() * 0.98, movement.getZ() * speedModifier);
-        
-        this.move(MovementType.SELF, this.getVelocity());
-    }
-
-    @Override
     public ActionResult interact(PlayerEntity player, Hand hand) {
         if (player.shouldCancelInteraction()) {
             return ActionResult.PASS;
         }
-        if (this.hasPassengers()) {
-            return ActionResult.PASS;
-        }
         if (!this.world.isClient) {
+            if (this.getPassengerList().size() > this.getMaxPassengers()) {
+                return ActionResult.PASS;
+            }
             player.setYaw(this.getYaw());
             player.setPitch(this.getPitch());
             return player.startRiding(this) ? ActionResult.CONSUME : ActionResult.PASS;
@@ -189,24 +208,45 @@ public class VehicleEntity extends Entity {
         return ActionResult.SUCCESS;
     }
 
+    public void openInventory(PlayerEntity player) {
+        openInventory(player, new VehicleScreenHandlerFactory(this));
+    }
+
+    public void openInventory(PlayerEntity player, ExtendedScreenHandlerFactory handler) {
+        if (!player.world.isClient) {
+            if (player.isSneaking()) {
+                player.openHandledScreen(handler);
+            }
+        }
+    }
+
     @Override
     public boolean damage(DamageSource source, float amount) {
         if (amount > 0) {
             if (source.getAttacker() instanceof PlayerEntity player) {
-                this.drop();
-                return true;
+                if (!(player.getVehicle() instanceof VehicleEntity)) {
+                    this.drop();
+                    return true;
+                }
             }
         }
         return false;
     }
 
     public void drop() {
-        if (getDropStack() != null) {
+        if (getDropStack() != null && this.world.getGameRules().getBoolean(GameRules.DO_TILE_DROPS)) {
             BlockPos pos = this.getBlockPos();
             ItemStack dropStack = this.getDropStack();
+
+            // Set the fluid and fluid variant in the dropped item.
+            ((VehicleItem) dropStack.getItem()).setAmount(dropStack, this.inputTank.amount);
+            ((VehicleItem) dropStack.getItem()).setFluid(dropStack, this.inputTank.variant);
+            NbtCompound nbt = dropStack.getOrCreateNbt();
+            // Set the inventory in the dropped item.
+            nbt.put("Inventory", this.inventory.toNbtList());
+
             world.playSound(null, pos, SoundEvents.BLOCK_NETHERITE_BLOCK_BREAK, SoundCategory.BLOCKS, 1, 1);
             this.world.spawnEntity(new ItemEntity(this.world, pos.getX(), pos.getY() + 0.5f, pos.getZ(), dropStack));
-            // TODO: Drop inventory and save nbt in item stack.
         }
 
         if (!this.world.isClient) {
@@ -216,37 +256,22 @@ public class VehicleEntity extends Entity {
 
     public void explode(float powerMultiplier) {
         if (!this.world.isClient) {
-            for (int i = 0; i < 3; i++) {
-                world.createExplosion(null, this.getX(), this.getY() + 0.5, this.getZ(), 7.0f * powerMultiplier, ModUtils.worldHasOxygen(this.world), Explosion.DestructionType.DESTROY);
-            }
+            world.createExplosion(this, this.getX(), this.getY() + 0.5, this.getZ(), 7.0f * powerMultiplier, ModUtils.worldHasOxygen(this.world), Explosion.DestructionType.DESTROY);
         }
         this.remove(RemovalReason.DISCARDED);
     }
 
     @Override
     public boolean handleFallDamage(float fallDistance, float damageMultiplier, DamageSource damageSource) {
-        if (this.getVelocity().getY() < -1.0) {
-            this.explode(1);
+        if (this.getVelocity().getY() < BeyondEarth.CONFIG.vehicles.fallingExplosionThreshold) {
+            this.explode(BeyondEarth.CONFIG.vehicles.fallingExplosionMultiplier);
+            return true;
         }
-        return super.handleFallDamage(fallDistance, damageMultiplier, damageSource);
+        return false;
     }
 
     public ItemStack getDropStack() {
         return null;
-    }
-
-    private float getMovementSpeed(float slipperiness) {
-        return (float) (this.getSpeed() * (0.21600002f / (slipperiness * slipperiness * slipperiness)));
-    }
-
-    public Vec3d applyMovementInput(Vec3d movementInput, float slipperiness) {
-        this.updateVelocity(this.getMovementSpeed(slipperiness), movementInput);
-        this.move(MovementType.SELF, this.getVelocity());
-        Vec3d vec3d = this.getVelocity();
-        if (this.horizontalCollision && (this.getBlockStateAtPos().isOf(Blocks.POWDER_SNOW) && PowderSnowBlock.canWalkOnPowderSnow(this))) {
-            vec3d = new Vec3d(vec3d.x, 0.2, vec3d.z);
-        }
-        return vec3d;
     }
 
     @Override
@@ -292,5 +317,60 @@ public class VehicleEntity extends Entity {
 
     public boolean renderPlanetBar() {
         return false;
+    }
+
+    @Override
+    public boolean collides() {
+        return true;
+    }
+
+    @Override
+    public boolean isCollidable() {
+        return true;
+    }
+
+    public long getTankSize() {
+        return 0;
+    }
+
+    public long getFuelPerTick() {
+        return 0;
+    }
+
+    public CustomInventory getInventory() {
+        return this.inventory;
+    }
+
+    public abstract int getInventorySize();
+
+    @Override
+    public Packet<?> createSpawnPacket() {
+        return new EntitySpawnS2CPacket(this);
+    }
+
+    public void tryInsertingIntoTank() {
+        if (this.getInventorySize() > 1) {
+            if (!this.world.isClient) {
+                FluidUtils.insertFluidIntoTank(this.getInventory(), this.inputTank, 0, 1, f -> f.equals(FluidVariant.of(ModFluids.FUEL_STILL)));
+            }
+        }
+    }
+
+    public int getFluidAmount() {
+        return this.dataTracker.get(FLUID_AMOUNT);
+    }
+
+    public FluidVariant getFluidVariant() {
+        return FluidVariant.of(Registry.FLUID.get(new Identifier(this.dataTracker.get(FLUID_VARIANT))));
+    }
+
+    public void consumeFuel() {
+        if (!this.world.isClient) {
+            try (Transaction transaction = Transaction.openOuter()) {
+                if (this.inputTank.extract(FluidVariant.of(ModFluids.FUEL_STILL), this.getFuelPerTick(), transaction) > 0) {
+                    transaction.commit();
+                }
+            }
+        }
     }
 }
